+++ ModLoaders/src/main/java/com/hypherionmc/sdlink/loaders/forge/SDLinkForge.java
@@ -5,6 +5,9 @@ import com.hypherionmc.craterlib.core.platform.ModloaderEnvironment;
 import com.hypherionmc.sdlink.SDLinkConstants;
 import com.hypherionmc.sdlink.client.ClientEvents;
 import com.hypherionmc.sdlink.compat.MModeCompat;
+import com.hypherionmc.sdlink.compat.rolesync.RoleSync;
+import com.hypherionmc.sdlink.compat.rolesync.impl.FTBRankSync;
+import com.hypherionmc.sdlink.compat.rolesync.impl.LuckPermsSync;
 import com.hypherionmc.sdlink.networking.SDLinkNetworking;
 import com.hypherionmc.sdlink.server.ServerEvents;
 import net.minecraftforge.api.distmarker.Dist;
@@ -23,6 +26,14 @@ public final class SDLinkForge {
 
             if (ModloaderEnvironment.INSTANCE.isModLoaded("mmode")) {
                 MModeCompat.init();
+            }
+
+            if (ModloaderEnvironment.INSTANCE.isModLoaded("ftbranks")) {
+                CraterEventBus.INSTANCE.registerEventListener(FTBRankSync.INSTANCE);
+            }
+
+            if (ModloaderEnvironment.INSTANCE.isModLoaded("luckperms")) {
+                CraterEventBus.INSTANCE.registerEventListener(LuckPermsSync.INSTANCE);
             }
         });