+++ ModLoaders/src/main/java/com/hypherionmc/sdlink/loaders/neoforge/SDLinkNeoForge.java
@@ -5,6 +5,9 @@ import com.hypherionmc.craterlib.core.platform.ModloaderEnvironment;
 import com.hypherionmc.sdlink.SDLinkConstants;
 import com.hypherionmc.sdlink.client.ClientEvents;
 import com.hypherionmc.sdlink.compat.MModeCompat;
+import com.hypherionmc.sdlink.compat.rolesync.RoleSync;
+import com.hypherionmc.sdlink.compat.rolesync.impl.FTBRankSync;
+import com.hypherionmc.sdlink.compat.rolesync.impl.LuckPermsSync;
 import com.hypherionmc.sdlink.networking.SDLinkNetworking;
 import com.hypherionmc.sdlink.server.ServerEvents;
 import net.neoforged.bus.api.IEventBus;
@@ -22,6 +25,14 @@ public final class SDLinkNeoForge {
 
             if (ModloaderEnvironment.INSTANCE.isModLoaded("mmode")) {
                 MModeCompat.init();
+            }
+
+            if (ModloaderEnvironment.INSTANCE.isModLoaded("ftbranks")) {
+                CraterEventBus.INSTANCE.registerEventListener(FTBRankSync.INSTANCE);
+            }
+
+            if (ModloaderEnvironment.INSTANCE.isModLoaded("luckperms")) {
+                CraterEventBus.INSTANCE.registerEventListener(LuckPermsSync.INSTANCE);
             }
         }