+++ changelog.md
@@ -25,5 +25,6 @@
 - Add option to trigger Minecraft commands on role changes [#77](https://github.com/hypherionmc/sdlink/issues/77)
 - Added a basic developer api. Please request additional features on GitHub/Discord
 - Added `sdconfigeditor` minecraft command, to integrate with our Online Config Editor
-- Support for Paper. Only currently available on Modrinth and NightBloom
+- Support for Paper. Only currently available on [NightBloom](https://nightbloom.cc/project/sdlink/files?loader=paper)
 - Removed reliance on JSONDB and greatly reduce mod size
+- FTB Ranks and LuckPerms ranks syncing